Spread receiver and sender across different virtual-networks
Sender is connected to one control-node and receiver is connected
to another. Hence the routes are learned over bgp in this scenario.

Fix a couple of issues found there by. In particular, when Type-4 path
needs to be replicated, we can identify the target table by extracting
the type-3 prefix from the type-4 prefix, and then get to the target
table using the src_table() of the associated imported secondary
type-3 path.

This is a combined work from Nikhil and Ananth

Change-Id: I668fe9b9fbd47e59490e6ce1850257aef0e8e7b0
Partial-Bug: #1708403

Conflicts:
	src/bgp/mvpn/mvpn_table.cc
	src/bgp/test/bgp_mvpn_integration_test.cc